style (auth): move NbAuthResult to separate file
