+++ docs/output.json
@@ -390,10 +390,10 @@
           "platform": null,
           "name": "isAuthenticated",
           "type": [
-            "Observable<any>"
+            "Observable<boolean>"
           ],
           "isStatic": false,
-          "shortDescription": "Returns true if auth token is presented in the token storage\n// TODO: check exp date for JWT token"
+          "shortDescription": "Returns true if auth token is presented in the token storage\n// TODO: check exp date for JWT token\n// TODO: for implement previous todo lets use isValid method of token"
         },
         {
           "examples": [],
@@ -422,7 +422,7 @@
             "Observable<boolean>"
           ],
           "isStatic": false,
-          "shortDescription": "Returns authentication status stream\n // TODO: check exp date for JWT token"
+          "shortDescription": "Returns authentication status stream\n // TODO: check exp date for JWT token\n // TODO: for implement previous todo lets use isValid method of token"
         },
         {
           "examples": [],
