+++ src/framework/auth/components/register/register.component.ts
@@ -8,7 +8,9 @@ import { Router } from '@angular/router';
 import { NB_AUTH_OPTIONS_TOKEN } from '../../auth.options';
 import { getDeepFromObject } from '../../helpers';
 
-import { NbAuthResult, NbAuthService } from '../../services/auth.service';
+import { NbAuthService } from '../../services/auth.service';
+import { NbAuthResult } from '../../services/auth-result';
+
 
 @Component({
   selector: 'nb-register',
