+++ src/framework/auth/providers/dummy-auth.provider.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { of as observableOf } from 'rxjs/observable/of';
 
-import { NbAuthResult } from '../services/auth.service';
+import { NbAuthResult } from '../services/auth-result';
 import { NbAbstractAuthProvider } from './abstract-auth.provider';
 import 'rxjs/add/operator/delay';
