+++ src/framework/auth/services/auth.service.ts
@@ -14,66 +14,7 @@ import { of as observableOf } from 'rxjs/observable/of';
 import { NbAbstractAuthProvider } from '../providers/abstract-auth.provider';
 import { NbAuthSimpleToken, NbTokenService } from './token.service';
 import { NB_AUTH_PROVIDERS_TOKEN } from '../auth.options';
-
-export class NbAuthResult {
-
-  protected token: any;
-  protected errors: string[] = [];
-  protected messages: string[] = [];
-
-  // TODO pass arguments in options object
-  constructor(protected success: boolean,
-    protected response?: any,
-    protected redirect?: any,
-    errors?: any,
-    messages?: any,
-    token?: NbAuthSimpleToken) {
-
-    this.errors = this.errors.concat([errors]);
-    if (errors instanceof Array) {
-      this.errors = errors;
-    }
-
-    this.messages = this.messages.concat([messages]);
-    if (messages instanceof Array) {
-      this.messages = messages;
-    }
-
-    this.token = token;
-  }
-
-  getResponse(): any {
-    return this.response;
-  }
-
-  getTokenValue(): any {
-    return this.token;
-  }
-
-  replaceToken(token: NbAuthSimpleToken): any {
-    this.token = token
-  }
-
-  getRedirect(): any {
-    return this.redirect;
-  }
-
-  getErrors(): string[] {
-    return this.errors.filter(val => !!val);
-  }
-
-  getMessages(): string[] {
-    return this.messages.filter(val => !!val);
-  }
-
-  isSuccess(): boolean {
-    return this.success;
-  }
-
-  isFailure(): boolean {
-    return !this.success;
-  }
-}
+import { NbAuthResult } from './auth-result';
 
 /**
  * Common authentication service.
